Add --gtirows control
Adds a --gtirows option to masterplotter.

This option takes a vector of integers representing a row selection in
the gti table. The indices run from 0 to len(gti)-1. NicerFileSet will
then apply only the selected gti rows.

If a custom gti table is given with --applygti, then the row selection
will operate on that custom table.