Updating cr_cut.py
In the cr_cut.py script, in step 5b, there is another print statement for the start and stop time of the last squeezed GTI added at the very end, outside of the for loop. Without this, the last set of GTIs would not be printed to the list of GTIs. One could miss anywhere from a bin size worth of data, to a few hundred seconds